CMS/Persistence/Repository/AuthorRepository.cs
using Model;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;
using NHibernate;
using Persistence.utils;

namespace Persistence.Repository
{
    class AuthorRepository : IRepository<Author>
    { 
        public void Save(Author person)
        {
            using (ISession session = NHibernateHelper.OpenSession())
            using (ITransaction transaction = session.BeginTransaction())
            {
                session.Save(person);
                transaction.Commit();
            }
        }
        public Author Get(int id)
        {
            using (ISession session = NHibernateHelper.OpenSession())
                return session.Get<Author>(id);
        }

        public void Update(Author person)
        {
            using (ISession session = NHibernateHelper.OpenSession())
            using (ITransaction transaction = session.BeginTransaction())
            {
                session.Update(person);
                transaction.Commit();
            }
        }

        public void Delete(Author person)
        {
            using (ISession session = NHibernateHelper.OpenSession())
            using (ITransaction transaction = session.BeginTransaction())
            {
                session.Delete(person);
                transaction.Commit();
            }
        }

        public long RowCount()
        {
            using (ISession session = NHibernateHelper.OpenSession())
            {
                return session.QueryOver<Author>().RowCountInt64();
            }
        }
    }
}

